+++ decidim-initiatives/app/services/decidim/initiatives/status_change_notifier.rb
@@ -58,6 +58,12 @@ module Decidim
       end
 
       def notify_support_result
+        Decidim::User.where(organization: initiative.organization, admin: true).each do |admin|
+          Decidim::Initiatives::InitiativesMailer
+            .notify_state_change(initiative, admin)
+            .deliver_later
+        end
+
         initiative.followers.each do |follower|
           Decidim::Initiatives::InitiativesMailer
             .notify_state_change(initiative, follower)